Update atari-v2.py
This implementation contains all of the components of the GoExplore algorithm that the prior one missed. It is far better in terms of convergence, ie. it discovers high reward trajectories far more rapidly than the prior implementation, which becomes especially evident as they are run for a longer duration of time.